Rename Vue components
