+++ webapp/spaceAdministration/src/main/webapp/vue-app/components/ExoSpacesAdministrationApp.vue
@@ -1,0 +1,13 @@
+<template>
+  <exo-spaces-administration-spaces></exo-spaces-administration-spaces>
+</template>
+
+<script>
+import ExoSpacesAdministrationSpaces from './ExoSpacesAdministrationSpaces.vue';
+
+export default {
+  components: {
+    'exo-spaces-administration-spaces': ExoSpacesAdministrationSpaces
+  }
+};
+</script>
