+++ webapp/spaceAdministration/src/main/webapp/vue-app/components/ExoSpacesAdministrationManageSpaces.vue
