+++ webapp/spaceAdministration/src/main/webapp/vue-app/components/ExoSpacesAdministrationSpacesPermissions.vue
