--- webapp/spaceAdministration/src/main/webapp/vue-app/components/app.vue
@@ -1,13 +1,0 @@
-<template>
-  <exo-space-administration-spaces></exo-space-administration-spaces>
-</template>
-
-<script>
-import ExoSpaceAdministrationSpaces from './ExoSpaceAdministrationSpaces.vue';
-
-export default {
-  components: {
-    'exo-space-administration-spaces': ExoSpaceAdministrationSpaces
-  }
-};
-</script>
